requirements.txt
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.1
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.8.0
torch==1.8.1+cu111
